third_party/blink/web_tests/platform/linux/virtual/idna-2008/external/wpt/url/a-element-xhtml_exclude=(file_javascript_mailto)-expected.txt

This is a testharness.js-based test.
Found 660 tests; 346 PASS, 314 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Loading data…
PASS Parsing: <http://example	.
org> against <http://example.org/foo/bar>
PASS Parsing: <http://user:pass@foo:21/bar;par?b#c> against <http://example.org/foo/bar>
PASS Parsing: <https://test:@test> against <about:blank>
PASS Parsing: <https://:@test> against <about:blank>
FAIL Parsing: <non-special://test:@test/x> against <about:blank> assert_equals: href expected "non-special://test@test/x" but got "non-special://test:@test/x"
FAIL Parsing: <non-special://:@test/x> against <about:blank> assert_equals: href expected "non-special://test/x" but got "non-special://:@test/x"
PASS Parsing: <http:foo.com> against <http://example.org/foo/bar>
PASS Parsing: <	   :foo.com   
> against <http://example.org/foo/bar>
PASS Parsing: < foo.com  > against <http://example.org/foo/bar>
PASS Parsing: <a:	 foo.com> against <http://example.org/foo/bar>
PASS Parsing: <http://f:21/ b ? d # e > against <http://example.org/foo/bar>
PASS Parsing: <lolscheme:x x#x x> against <about:blank>
PASS Parsing: <http://f:/c> against <http://example.org/foo/bar>
PASS Parsing: <http://f:0/c> against <http://example.org/foo/bar>
PASS Parsing: <http://f:00000000000000/c> against <http://example.org/foo/bar>
PASS Parsing: <http://f:00000000000000000000080/c> against <http://example.org/foo/bar>
PASS Parsing: <http://f:b/c> against <http://example.org/foo/bar>
FAIL Parsing: <http://f: /c> against <http://example.org/foo/bar> assert_equals: failure should set href to input expected "http://f: /c" but got "http://f:%20/c"
PASS Parsing: <http://f:
/c> against <http://example.org/foo/bar>
PASS Parsing: <http://f:fifty-two/c> against <http://example.org/foo/bar>
PASS Parsing: <http://f:999999/c> against <http://example.org/foo/bar>
FAIL Parsing: <non-special://f:999999/c> against <http://example.org/foo/bar> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://f: 21 / b ? d # e > against <http://example.org/foo/bar> assert_equals: failure should set href to input expected "http://f: 21 / b ? d # e " but got "http://f:%2021%20/%20b%20?%20d%20#%20e"
PASS Parsing: <> against <http://example.org/foo/bar>
PASS Parsing: <  	> against <http://example.org/foo/bar>
PASS Parsing: <:foo.com/> against <http://example.org/foo/bar>
PASS Parsing: <:foo.com\> against <http://example.org/foo/bar>
PASS Parsing: <:> against <http://example.org/foo/bar>
PASS Parsing: <:a> against <http://example.org/foo/bar>
PASS Parsing: <:/> against <http://example.org/foo/bar>
PASS Parsing: <:\> against <http://example.org/foo/bar>
PASS Parsing: <:#> against <http://example.org/foo/bar>
PASS Parsing: <#> against <http://example.org/foo/bar>
PASS Parsing: <#/> against <http://example.org/foo/bar>
PASS Parsing: <#\> against <http://example.org/foo/bar>
PASS Parsing: <#;?> against <http://example.org/foo/bar>
PASS Parsing: <?> against <http://example.org/foo/bar>
PASS Parsing: </> against <http://example.org/foo/bar>
PASS Parsing: <:23> against <http://example.org/foo/bar>
PASS Parsing: </:23> against <http://example.org/foo/bar>
PASS Parsing: <\x> against <http://example.org/foo/bar>
PASS Parsing: <\\x\hello> against <http://example.org/foo/bar>
PASS Parsing: <::> against <http://example.org/foo/bar>
PASS Parsing: <::23> against <http://example.org/foo/bar>
FAIL Parsing: <foo://> against <http://example.org/foo/bar> assert_equals: pathname expected "" but got "//"
PASS Parsing: <http://a:b@c:29/d> against <http://example.org/foo/bar>
PASS Parsing: <http::@c:29> against <http://example.org/foo/bar>
PASS Parsing: <http://&a:foo(b]c@d:2/> against <http://example.org/foo/bar>
PASS Parsing: <http://::@c@d:2> against <http://example.org/foo/bar>
PASS Parsing: <http://foo.com:b@d/> against <http://example.org/foo/bar>
PASS Parsing: <http://foo.com/\@> against <http://example.org/foo/bar>
PASS Parsing: <http:\\foo.com\> against <http://example.org/foo/bar>
PASS Parsing: <http:\\a\b:c\d@foo.com\> against <http://example.org/foo/bar>
PASS Parsing: <foo:/> against <http://example.org/foo/bar>
PASS Parsing: <foo:/bar.com/> against <http://example.org/foo/bar>
FAIL Parsing: <foo://///////> against <http://example.org/foo/bar> assert_equals: pathname expected "///////" but got "/////////"
FAIL Parsing: <foo://///////bar.com/> against <http://example.org/foo/bar> assert_equals: pathname expected "///////bar.com/" but got "/////////bar.com/"
FAIL Parsing: <foo:////://///> against <http://example.org/foo/bar> assert_equals: pathname expected "//://///" but got "////://///"
PASS Parsing: <c:/foo> against <http://example.org/foo/bar>
PASS Parsing: <//foo/bar> against <http://example.org/foo/bar>
PASS Parsing: <http://foo/path;a??e#f#g> against <http://example.org/foo/bar>
PASS Parsing: <http://foo/abcd?efgh?ijkl> against <http://example.org/foo/bar>
PASS Parsing: <http://foo/abcd#foo?bar> against <http://example.org/foo/bar>
PASS Parsing: <[61:24:74]:98> against <http://example.org/foo/bar>
PASS Parsing: <http:[61:27]/:foo> against <http://example.org/foo/bar>
FAIL Parsing: <http://[1::2]:3:4> against <http://example.org/foo/bar> assert_equals: failure should set href to input expected "http://[1::2]:3:4" but got "http://[1::2]:3:4/"
FAIL Parsing: <http://2001::1> against <http://example.org/foo/bar> assert_equals: failure should set href to input expected "http://2001::1" but got "http://2001::1/"
FAIL Parsing: <http://2001::1]> against <http://example.org/foo/bar> assert_equals: failure should set href to input expected "http://2001::1]" but got "http://2001::1]/"
FAIL Parsing: <http://2001::1]:80> against <http://example.org/foo/bar> assert_equals: failure should set href to input expected "http://2001::1]:80" but got "http://2001::1]/"
PASS Parsing: <http://[2001::1]> against <http://example.org/foo/bar>
PASS Parsing: <http://[::127.0.0.1]> against <http://example.org/foo/bar>
FAIL Parsing: <http://[::127.0.0.1.]> against <http://example.org/foo/bar> assert_unreached: Expected URL to fail parsing Reached unreachable code
PASS Parsing: <http://[0:0:0:0:0:0:13.1.68.3]> against <http://example.org/foo/bar>
PASS Parsing: <http://[2001::1]:80> against <http://example.org/foo/bar>
PASS Parsing: <http:/example.com/> against <http://example.org/foo/bar>
PASS Parsing: <ftp:/example.com/> against <http://example.org/foo/bar>
PASS Parsing: <https:/example.com/> against <http://example.org/foo/bar>
PASS Parsing: <madeupscheme:/example.com/> against <http://example.org/foo/bar>
PASS Parsing: <ftps:/example.com/> against <http://example.org/foo/bar>
PASS Parsing: <gopher:/example.com/> against <http://example.org/foo/bar>
PASS Parsing: <ws:/example.com/> against <http://example.org/foo/bar>
PASS Parsing: <wss:/example.com/> against <http://example.org/foo/bar>
PASS Parsing: <data:/example.com/> against <http://example.org/foo/bar>
PASS Parsing: <http:example.com/> against <http://example.org/foo/bar>
PASS Parsing: <ftp:example.com/> against <http://example.org/foo/bar>
PASS Parsing: <https:example.com/> against <http://example.org/foo/bar>
PASS Parsing: <madeupscheme:example.com/> against <http://example.org/foo/bar>
PASS Parsing: <ftps:example.com/> against <http://example.org/foo/bar>
PASS Parsing: <gopher:example.com/> against <http://example.org/foo/bar>
PASS Parsing: <ws:example.com/> against <http://example.org/foo/bar>
PASS Parsing: <wss:example.com/> against <http://example.org/foo/bar>
PASS Parsing: <data:example.com/> against <http://example.org/foo/bar>
PASS Parsing: </a/b/c> against <http://example.org/foo/bar>
PASS Parsing: </a/ /c> against <http://example.org/foo/bar>
PASS Parsing: </a%2fc> against <http://example.org/foo/bar>
PASS Parsing: </a/%2f/c> against <http://example.org/foo/bar>
PASS Parsing: <#β> against <http://example.org/foo/bar>
PASS Parsing: <data:text/html,test#test> against <http://example.org/foo/bar>
PASS Parsing: <tel:1234567890> against <http://example.org/foo/bar>
FAIL Parsing: <ssh://example.com/foo/bar.git> against <http://example.org/> assert_equals: host expected "example.com" but got ""
PASS Parsing: <http://example.com/././foo> against <about:blank>
PASS Parsing: <http://example.com/./.foo> against <about:blank>
PASS Parsing: <http://example.com/foo/.> against <about:blank>
PASS Parsing: <http://example.com/foo/./> against <about:blank>
PASS Parsing: <http://example.com/foo/bar/..> against <about:blank>
PASS Parsing: <http://example.com/foo/bar/../> against <about:blank>
PASS Parsing: <http://example.com/foo/..bar> against <about:blank>
PASS Parsing: <http://example.com/foo/bar/../ton> against <about:blank>
PASS Parsing: <http://example.com/foo/bar/../ton/../../a> against <about:blank>
PASS Parsing: <http://example.com/foo/../../..> against <about:blank>
PASS Parsing: <http://example.com/foo/../../../ton> against <about:blank>
PASS Parsing: <http://example.com/foo/%2e> against <about:blank>
FAIL Parsing: <http://example.com/foo/%2e%2> against <about:blank> assert_equals: href expected "http://example.com/foo/%2e%2" but got "http://example.com/foo/.%2"
FAIL Parsing: <http://example.com/foo/%2e./%2e%2e/.%2e/%2e.bar> against <about:blank> assert_equals: href expected "http://example.com/%2e.bar" but got "http://example.com/..bar"
PASS Parsing: <http://example.com////../..> against <about:blank>
PASS Parsing: <http://example.com/foo/bar//../..> against <about:blank>
PASS Parsing: <http://example.com/foo/bar//..> against <about:blank>
PASS Parsing: <http://example.com/foo> against <about:blank>
PASS Parsing: <http://example.com/%20foo> against <about:blank>
PASS Parsing: <http://example.com/foo%> against <about:blank>
PASS Parsing: <http://example.com/foo%2> against <about:blank>
PASS Parsing: <http://example.com/foo%2zbar> against <about:blank>
PASS Parsing: <http://example.com/foo%2Â©zbar> against <about:blank>
FAIL Parsing: <http://example.com/foo%41%7a> against <about:blank> assert_equals: href expected "http://example.com/foo%41%7a" but got "http://example.com/fooAz"
PASS Parsing: <http://example.com/foo	%91> against <about:blank>
FAIL Parsing: <http://example.com/foo%00%51> against <about:blank> assert_equals: href expected "http://example.com/foo%00%51" but got "http://example.com/foo%00Q"
PASS Parsing: <http://example.com/(%28:%3A%29)> against <about:blank>
PASS Parsing: <http://example.com/%3A%3a%3C%3c> against <about:blank>
PASS Parsing: <http://example.com/foo	bar> against <about:blank>
PASS Parsing: <http://example.com\\foo\\bar> against <about:blank>
PASS Parsing: <http://example.com/%7Ffp3%3Eju%3Dduvgw%3Dd> against <about:blank>
PASS Parsing: <http://example.com/@asdf%40> against <about:blank>
PASS Parsing: <http://example.com/你好你好> against <about:blank>
PASS Parsing: <http://example.com/‥/foo> against <about:blank>
PASS Parsing: <http://example.com/﻿/foo> against <about:blank>
PASS Parsing: <http://example.com/‮/foo/‭/bar> against <about:blank>
PASS Parsing: <http://www.google.com/foo?bar=baz#> against <about:blank>
PASS Parsing: <http://www.google.com/foo?bar=baz# »> against <about:blank>
PASS Parsing: <data:test# »> against <about:blank>
PASS Parsing: <http://www.google.com> against <about:blank>
PASS Parsing: <http://192.0x00A80001> against <about:blank>
FAIL Parsing: <http://www/foo%2Ehtml> against <about:blank> assert_equals: href expected "http://www/foo%2Ehtml" but got "http://www/foo.html"
PASS Parsing: <http://www/foo/%2E/html> against <about:blank>
PASS Parsing: <http://user:pass@/> against <about:blank>
PASS Parsing: <http://%25DOMAIN:foobar@foodomain.com/> against <about:blank>
PASS Parsing: <http:\\www.google.com\foo> against <about:blank>
PASS Parsing: <http://foo:80/> against <about:blank>
PASS Parsing: <http://foo:81/> against <about:blank>
FAIL Parsing: <httpa://foo:80/> against <about:blank> assert_equals: host expected "foo:80" but got ""
PASS Parsing: <http://foo:-80/> against <about:blank>
PASS Parsing: <https://foo:443/> against <about:blank>
PASS Parsing: <https://foo:80/> against <about:blank>
PASS Parsing: <ftp://foo:21/> against <about:blank>
PASS Parsing: <ftp://foo:80/> against <about:blank>
FAIL Parsing: <gopher://foo:70/> against <about:blank> assert_equals: host expected "foo:70" but got ""
FAIL Parsing: <gopher://foo:443/> against <about:blank> assert_equals: host expected "foo:443" but got ""
PASS Parsing: <ws://foo:80/> against <about:blank>
PASS Parsing: <ws://foo:81/> against <about:blank>
PASS Parsing: <ws://foo:443/> against <about:blank>
PASS Parsing: <ws://foo:815/> against <about:blank>
PASS Parsing: <wss://foo:80/> against <about:blank>
PASS Parsing: <wss://foo:81/> against <about:blank>
PASS Parsing: <wss://foo:443/> against <about:blank>
PASS Parsing: <wss://foo:815/> against <about:blank>
PASS Parsing: <http:/example.com/> against <about:blank>
PASS Parsing: <ftp:/example.com/> against <about:blank>
PASS Parsing: <https:/example.com/> against <about:blank>
PASS Parsing: <madeupscheme:/example.com/> against <about:blank>
PASS Parsing: <ftps:/example.com/> against <about:blank>
PASS Parsing: <gopher:/example.com/> against <about:blank>
PASS Parsing: <ws:/example.com/> against <about:blank>
PASS Parsing: <wss:/example.com/> against <about:blank>
PASS Parsing: <data:/example.com/> against <about:blank>
PASS Parsing: <http:example.com/> against <about:blank>
PASS Parsing: <ftp:example.com/> against <about:blank>
PASS Parsing: <https:example.com/> against <about:blank>
PASS Parsing: <madeupscheme:example.com/> against <about:blank>
PASS Parsing: <ftps:example.com/> against <about:blank>
PASS Parsing: <gopher:example.com/> against <about:blank>
PASS Parsing: <ws:example.com/> against <about:blank>
PASS Parsing: <wss:example.com/> against <about:blank>
PASS Parsing: <data:example.com/> against <about:blank>
PASS Parsing: <http:@www.example.com> against <about:blank>
PASS Parsing: <http:/@www.example.com> against <about:blank>
PASS Parsing: <http://@www.example.com> against <about:blank>
PASS Parsing: <http:a:b@www.example.com> against <about:blank>
PASS Parsing: <http:/a:b@www.example.com> against <about:blank>
PASS Parsing: <http://a:b@www.example.com> against <about:blank>
PASS Parsing: <http://@pple.com> against <about:blank>
PASS Parsing: <http::b@www.example.com> against <about:blank>
PASS Parsing: <http:/:b@www.example.com> against <about:blank>
PASS Parsing: <http://:b@www.example.com> against <about:blank>
FAIL Parsing: <http:/:@/www.example.com> against <about:blank> assert_equals: failure should set href to input expected "http:/:@/www.example.com" but got "http:///www.example.com"
PASS Parsing: <http://user@/www.example.com> against <about:blank>
FAIL Parsing: <http:@/www.example.com> against <about:blank> assert_equals: failure should set href to input expected "http:@/www.example.com" but got "http:///www.example.com"
FAIL Parsing: <http:/@/www.example.com> against <about:blank> assert_equals: failure should set href to input expected "http:/@/www.example.com" but got "http:///www.example.com"
FAIL Parsing: <http://@/www.example.com> against <about:blank> assert_equals: failure should set href to input expected "http://@/www.example.com" but got "http:///www.example.com"
FAIL Parsing: <https:@/www.example.com> against <about:blank> assert_equals: failure should set href to input expected "https:@/www.example.com" but got "https:///www.example.com"
FAIL Parsing: <http:a:b@/www.example.com> against <about:blank> assert_equals: failure should set href to input expected "http:a:b@/www.example.com" but got "http://a:b@/www.example.com"
FAIL Parsing: <http:/a:b@/www.example.com> against <about:blank> assert_equals: failure should set href to input expected "http:/a:b@/www.example.com" but got "http://a:b@/www.example.com"
PASS Parsing: <http://a:b@/www.example.com> against <about:blank>
FAIL Parsing: <http::@/www.example.com> against <about:blank> assert_equals: failure should set href to input expected "http::@/www.example.com" but got "http:///www.example.com"
PASS Parsing: <http:a:@www.example.com> against <about:blank>
PASS Parsing: <http:/a:@www.example.com> against <about:blank>
PASS Parsing: <http://a:@www.example.com> against <about:blank>
PASS Parsing: <http://www.@pple.com> against <about:blank>
FAIL Parsing: <http:@:www.example.com> against <about:blank> assert_equals: failure should set href to input expected "http:@:www.example.com" but got "http://:www.example.com/"
FAIL Parsing: <http:/@:www.example.com> against <about:blank> assert_equals: failure should set href to input expected "http:/@:www.example.com" but got "http://:www.example.com/"
FAIL Parsing: <http://@:www.example.com> against <about:blank> assert_equals: failure should set href to input expected "http://@:www.example.com" but got "http://:www.example.com/"
PASS Parsing: <http://:@www.example.com> against <about:blank>
PASS Parsing: </> against <http://www.example.com/test>
PASS Parsing: </test.txt> against <http://www.example.com/test>
PASS Parsing: <.> against <http://www.example.com/test>
PASS Parsing: <..> against <http://www.example.com/test>
PASS Parsing: <test.txt> against <http://www.example.com/test>
PASS Parsing: <./test.txt> against <http://www.example.com/test>
PASS Parsing: <../test.txt> against <http://www.example.com/test>
PASS Parsing: <../aaa/test.txt> against <http://www.example.com/test>
PASS Parsing: <../../test.txt> against <http://www.example.com/test>
PASS Parsing: <中/test.txt> against <http://www.example.com/test>
PASS Parsing: <http://www.example2.com> against <http://www.example.com/test>
PASS Parsing: <//www.example2.com> against <http://www.example.com/test>
PASS Parsing: <http://ExAmPlE.CoM> against <http://other.com/>
FAIL Parsing: <http://example example.com> against <http://other.com/> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://Goo%20 goo%7C|.com> against <http://other.com/> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://[]> against <http://other.com/> assert_equals: failure should set href to input expected "http://[]" but got "http://[]/"
FAIL Parsing: <http://[:]> against <http://other.com/> assert_equals: failure should set href to input expected "http://[:]" but got "http://[:]/"
FAIL Parsing: <http://GOO 　goo.com> against <http://other.com/> assert_unreached: Expected URL to fail parsing Reached unreachable code
PASS Parsing: <http://GOO​⁠﻿goo.com> against <http://other.com/>
PASS Parsing: <\0 http://example.com/ \r > against <about:blank>
PASS Parsing: <http://www.foo。bar.com> against <http://other.com/>
FAIL Parsing: <http://﷐zyx.com> against <http://other.com/> assert_equals: failure should set href to input expected "http://﷐zyx.com" but got "http://%EF%BF%BDzyx.com/"
FAIL Parsing: <http://%ef%b7%90zyx.com> against <http://other.com/> assert_equals: failure should set href to input expected "http://%ef%b7%90zyx.com" but got "http://%EF%BF%BDzyx.com/"
FAIL Parsing: <https://�> against <about:blank> assert_equals: failure should set href to input expected "https://\ufffd" but got "https://%EF%BF%BD/"
FAIL Parsing: <https://%EF%BF%BD> against <about:blank> assert_equals: failure should set href to input expected "https://%EF%BF%BD" but got "https://%EF%BF%BD/"
PASS Parsing: <https://x/�?�#�> against <about:blank>
FAIL Parsing: <http://a.b.c.xn--pokxncvks> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://10.0.0.xn--pokxncvks> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://a.b.c.XN--pokxncvks> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://a.b.c.Xn--pokxncvks> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://10.0.0.XN--pokxncvks> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://10.0.0.xN--pokxncvks> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
PASS Parsing: <http://Ｇｏ.com> against <http://other.com/>
FAIL Parsing: <http://％４１.com> against <http://other.com/> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://%ef%bc%85%ef%bc%94%ef%bc%91.com> against <http://other.com/> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://％００.com> against <http://other.com/> assert_equals: failure should set href to input expected "http://％００.com" but got "http://%00.com/"
FAIL Parsing: <http://%ef%bc%85%ef%bc%90%ef%bc%90.com> against <http://other.com/> assert_equals: failure should set href to input expected "http://%ef%bc%85%ef%bc%90%ef%bc%90.com" but got "http://%00.com/"
PASS Parsing: <http://你好你好> against <http://other.com/>
FAIL Parsing: <https://faß.ExAmPlE/> against <about:blank> assert_equals: href expected "https://xn--fa-hia.example/" but got "https://fass.example/"
FAIL Parsing: <sc://faß.ExAmPlE/> against <about:blank> assert_equals: host expected "fa%C3%9F.ExAmPlE" but got ""
FAIL Parsing: <http://%zz%66%a.com> against <http://other.com/> assert_equals: failure should set href to input expected "http://%zz%66%a.com" but got "http://%25zzf%25a.com/"
FAIL Parsing: <http://%25> against <http://other.com/> assert_equals: failure should set href to input expected "http://%25" but got "http://%25/"
FAIL Parsing: <http://hello%00> against <http://other.com/> assert_equals: failure should set href to input expected "http://hello%00" but got "http://hello%00/"
PASS Parsing: <http://%30%78%63%30%2e%30%32%35%30.01> against <http://other.com/>
PASS Parsing: <http://%30%78%63%30%2e%30%32%35%30.01%2e> against <http://other.com/>
FAIL Parsing: <http://192.168.0.257> against <http://other.com/> assert_equals: failure should set href to input expected "http://192.168.0.257" but got "http://192.168.0.257/"
FAIL Parsing: <http://%3g%78%63%30%2e%30%32%35%30%2E.01> against <http://other.com/> assert_equals: failure should set href to input expected "http://%3g%78%63%30%2e%30%32%35%30%2E.01" but got "http://%253gxc0.0250..01/"
FAIL Parsing: <http://192.168.0.1 hello> against <http://other.com/> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <https://x x:12> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
PASS Parsing: <http://０Ｘｃ０．０２５０．０１> against <http://other.com/>
PASS Parsing: <http://./> against <about:blank>
PASS Parsing: <http://../> against <about:blank>
FAIL Parsing: <h://.> against <about:blank> assert_equals: host expected "." but got ""
PASS Parsing: <http://[www.google.com]/> against <about:blank>
FAIL Parsing: <http://[google.com]> against <http://other.com/> assert_equals: failure should set href to input expected "http://[google.com]" but got "http://[google.com]/"
FAIL Parsing: <http://[::1.2.3.4x]> against <http://other.com/> assert_equals: failure should set href to input expected "http://[::1.2.3.4x]" but got "http://[::1.2.3.4x]/"
FAIL Parsing: <http://[::1.2.3.]> against <http://other.com/> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://[::1.2.]> against <http://other.com/> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://[::.1.2]> against <http://other.com/> assert_equals: failure should set href to input expected "http://[::.1.2]" but got "http://[::.1.2]/"
FAIL Parsing: <http://[::1.]> against <http://other.com/> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://[::.1]> against <http://other.com/> assert_equals: failure should set href to input expected "http://[::.1]" but got "http://[::.1]/"
FAIL Parsing: <http://[::%31]> against <http://other.com/> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://%5B::1]> against <http://other.com/> assert_unreached: Expected URL to fail parsing Reached unreachable code
PASS Parsing: <http://foo:💩@example.com/bar> against <http://other.com/>
PASS Parsing: <#> against <test:test>
FAIL Parsing: <#x> against <data:,> assert_equals: href expected "data:,#x" but got "test:test#x"
PASS Parsing: <#x> against <about:blank>
PASS Parsing: <#> against <test:test?test>
PASS Parsing: <https://@test@test@example:800/> against <http://doesnotmatter/>
PASS Parsing: <https://@@@example> against <http://doesnotmatter/>
PASS Parsing: <http://`{}:`{}@h/`{}?`{}> against <http://doesnotmatter/>
PASS Parsing: <http://host/?'> against <about:blank>
FAIL Parsing: <notspecial://host/?'> against <about:blank> assert_equals: href expected "notspecial://host/?'" but got "notspecial://host/?%27"
PASS Parsing: </some/path> against <http://user@example.org/smth>
PASS Parsing: <> against <http://user:pass@example.org:21/smth>
PASS Parsing: </some/path> against <http://user:pass@example.org:21/smth>
FAIL Parsing: <i> against <sc:sd> assert_equals: failure should set href to input expected "i" but got ""
FAIL Parsing: <i> against <sc:sd/sd> assert_equals: failure should set href to input expected "i" but got ""
PASS Parsing: <i> against <sc:/pa/pa>
FAIL Parsing: <i> against <sc://ho/pa> assert_equals: host expected "ho" but got ""
FAIL Parsing: <i> against <sc:///pa/pa> assert_equals: pathname expected "/pa/i" but got "///pa/i"
FAIL Parsing: <../i> against <sc:sd> assert_equals: failure should set href to input expected "../i" but got ""
FAIL Parsing: <../i> against <sc:sd/sd> assert_equals: failure should set href to input expected "../i" but got ""
PASS Parsing: <../i> against <sc:/pa/pa>
FAIL Parsing: <../i> against <sc://ho/pa> assert_equals: host expected "ho" but got ""
FAIL Parsing: <../i> against <sc:///pa/pa> assert_equals: href expected "sc:///i" but got "sc:///pa/i"
FAIL Parsing: </i> against <sc:sd> assert_equals: failure should set href to input expected "/i" but got ""
FAIL Parsing: </i> against <sc:sd/sd> assert_equals: failure should set href to input expected "/i" but got ""
PASS Parsing: </i> against <sc:/pa/pa>
FAIL Parsing: </i> against <sc://ho/pa> assert_equals: host expected "ho" but got ""
FAIL Parsing: </i> against <sc:///pa/pa> assert_equals: href expected "sc:///i" but got "sc:///pa/i"
FAIL Parsing: <?i> against <sc:sd> assert_equals: failure should set href to input expected "?i" but got ""
FAIL Parsing: <?i> against <sc:sd/sd> assert_equals: failure should set href to input expected "?i" but got ""
PASS Parsing: <?i> against <sc:/pa/pa>
FAIL Parsing: <?i> against <sc://ho/pa> assert_equals: host expected "ho" but got ""
FAIL Parsing: <?i> against <sc:///pa/pa> assert_equals: pathname expected "/pa/pa" but got "///pa/pa"
PASS Parsing: <#i> against <sc:sd>
PASS Parsing: <#i> against <sc:sd/sd>
PASS Parsing: <#i> against <sc:/pa/pa>
FAIL Parsing: <#i> against <sc://ho/pa> assert_equals: host expected "ho" but got ""
FAIL Parsing: <#i> against <sc:///pa/pa> assert_equals: pathname expected "/pa/pa" but got "///pa/pa"
FAIL Parsing: <about:/../> against <about:blank> assert_equals: href expected "about:/" but got "about:/../"
FAIL Parsing: <data:/../> against <about:blank> assert_equals: href expected "data:/" but got "data:/../"
FAIL Parsing: <sc://ñ.test/> against <about:blank> assert_equals: host expected "%C3%B1.test" but got ""
FAIL Parsing: <sc://%/> against <about:blank> assert_equals: host expected "%" but got ""
FAIL Parsing: <sc://@/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <sc://te@s:t@/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <sc://:/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <sc://:12/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <x> against <sc://ñ> assert_equals: href expected "sc://%C3%B1/x" but got "sc://%C3%B1"
PASS Parsing: <sc:\../> against <about:blank>
PASS Parsing: <sc::a@example.net> against <about:blank>
PASS Parsing: <wow:%NBD> against <about:blank>
PASS Parsing: <wow:%1G> against <about:blank>
FAIL Parsing: <wow:￿> against <about:blank> assert_equals: href expected "wow:%EF%BF%BF" but got "wow:%EF%BF%BD"
FAIL Parsing: <http://example.com/U+d800𐟾U+dfff﷐﷏﷯ﷰ￾￿?U+d800𐟾U+dfff﷐﷏﷯ﷰ￾￿> against <about:blank> assert_equals: href expected "http://example.com/%EF%BF%BD%F0%90%9F%BE%EF%BF%BD%EF%B7%90%EF%B7%8F%EF%B7%AF%EF%B7%B0%EF%BF%BE%EF%BF%BF?%EF%BF%BD%F0%90%9F%BE%EF%BF%BD%EF%B7%90%EF%B7%8F%EF%B7%AF%EF%B7%B0%EF%BF%BE%EF%BF%BF" but got "http://example.com/%EF%BF%BD%F0%90%9F%BE%EF%BF%BD%EF%BF%BD%EF%B7%8F%EF%BF%BD%EF%B7%B0%EF%BF%BD%EF%BF%BD?%EF%BF%BD%F0%90%9F%BE%EF%BF%BD%EF%BF%BD%EF%B7%8F%EF%BF%BD%EF%B7%B0%EF%BF%BD%EF%BF%BD"
FAIL Parsing: <sc://a\0b/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <sc://a b/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <sc://a<b> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <sc://a>b> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <sc://a[b/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <sc://a\b/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <sc://a]b/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <sc://a^b> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <sc://a|b/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <foo://ho	st/> against <about:blank> assert_equals: host expected "host" but got ""
FAIL Parsing: <foo://ho
st/> against <about:blank> assert_equals: host expected "host" but got ""
FAIL Parsing: <foo://ho\rst/> against <about:blank> assert_equals: host expected "host" but got ""
FAIL Parsing: <http://a\0b/> against <about:blank> assert_equals: failure should set href to input expected "http://a\0b/" but got "http://a%00b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x01b/" but got "http://a%01b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x02b/" but got "http://a%02b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x03b/" but got "http://a%03b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x04b/" but got "http://a%04b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x05b/" but got "http://a%05b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x06b/" but got "http://a%06b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x07b/" but got "http://a%07b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\bb/" but got "http://a%08b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\vb/" but got "http://a%0Bb/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\fb/" but got "http://a%0Cb/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x0eb/" but got "http://a%0Eb/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x0fb/" but got "http://a%0Fb/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x10b/" but got "http://a%10b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x11b/" but got "http://a%11b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x12b/" but got "http://a%12b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x13b/" but got "http://a%13b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x14b/" but got "http://a%14b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x15b/" but got "http://a%15b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x16b/" but got "http://a%16b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x17b/" but got "http://a%17b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x18b/" but got "http://a%18b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x19b/" but got "http://a%19b/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x1ab/" but got "http://a%1Ab/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x1bb/" but got "http://a%1Bb/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x1cb/" but got "http://a%1Cb/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x1db/" but got "http://a%1Db/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x1eb/" but got "http://a%1Eb/"
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://a\x1fb/" but got "http://a%1Fb/"
FAIL Parsing: <http://a b/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://a%b/> against <about:blank> assert_equals: failure should set href to input expected "http://a%b/" but got "http://a%25b/"
FAIL Parsing: <http://a<b> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://a>b> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
PASS Parsing: <http://a[b/> against <about:blank>
PASS Parsing: <http://a]b/> against <about:blank>
FAIL Parsing: <http://a^b> against <about:blank> assert_equals: failure should set href to input expected "http://a^b" but got "http://a%5Eb/"
FAIL Parsing: <http://a|b/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://ab/> against <about:blank> assert_equals: failure should set href to input expected "http://ab/" but got "http://a%7Fb/"
PASS Parsing: <http://ho	st/> against <about:blank>
PASS Parsing: <http://ho
st/> against <about:blank>
PASS Parsing: <http://ho\rst/> against <about:blank>
PASS Parsing: <http://ho%00st/> against <about:blank>
PASS Parsing: <http://ho%01st/> against <about:blank>
PASS Parsing: <http://ho%02st/> against <about:blank>
PASS Parsing: <http://ho%03st/> against <about:blank>
PASS Parsing: <http://ho%04st/> against <about:blank>
PASS Parsing: <http://ho%05st/> against <about:blank>
PASS Parsing: <http://ho%06st/> against <about:blank>
PASS Parsing: <http://ho%07st/> against <about:blank>
PASS Parsing: <http://ho%08st/> against <about:blank>
PASS Parsing: <http://ho%09st/> against <about:blank>
PASS Parsing: <http://ho%0Ast/> against <about:blank>
PASS Parsing: <http://ho%0Bst/> against <about:blank>
PASS Parsing: <http://ho%0Cst/> against <about:blank>
PASS Parsing: <http://ho%0Dst/> against <about:blank>
PASS Parsing: <http://ho%0Est/> against <about:blank>
PASS Parsing: <http://ho%0Fst/> against <about:blank>
PASS Parsing: <http://ho%10st/> against <about:blank>
PASS Parsing: <http://ho%11st/> against <about:blank>
PASS Parsing: <http://ho%12st/> against <about:blank>
PASS Parsing: <http://ho%13st/> against <about:blank>
PASS Parsing: <http://ho%14st/> against <about:blank>
PASS Parsing: <http://ho%15st/> against <about:blank>
PASS Parsing: <http://ho%16st/> against <about:blank>
PASS Parsing: <http://ho%17st/> against <about:blank>
PASS Parsing: <http://ho%18st/> against <about:blank>
PASS Parsing: <http://ho%19st/> against <about:blank>
PASS Parsing: <http://ho%1Ast/> against <about:blank>
PASS Parsing: <http://ho%1Bst/> against <about:blank>
PASS Parsing: <http://ho%1Cst/> against <about:blank>
PASS Parsing: <http://ho%1Dst/> against <about:blank>
PASS Parsing: <http://ho%1Est/> against <about:blank>
PASS Parsing: <http://ho%1Fst/> against <about:blank>
FAIL Parsing: <http://ho%20st/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://ho%23st/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
PASS Parsing: <http://ho%25st/> against <about:blank>
PASS Parsing: <http://ho%2Fst/> against <about:blank>
FAIL Parsing: <http://ho%3Ast/> against <about:blank> assert_equals: failure should set href to input expected "http://ho%3Ast/" but got "http://ho:st/"
FAIL Parsing: <http://ho%3Cst/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://ho%3Est/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
PASS Parsing: <http://ho%3Fst/> against <about:blank>
FAIL Parsing: <http://ho%40st/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <http://ho%5Bst/> against <about:blank> assert_equals: failure should set href to input expected "http://ho%5Bst/" but got "http://ho[st/"
PASS Parsing: <http://ho%5Cst/> against <about:blank>
FAIL Parsing: <http://ho%5Dst/> against <about:blank> assert_equals: failure should set href to input expected "http://ho%5Dst/" but got "http://ho]st/"
FAIL Parsing: <http://ho%7Cst/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
PASS Parsing: <http://ho%7Fst/> against <about:blank>
FAIL Parsing: <http://!"$&'()*+,-.;=_`{}~/> against <about:blank> assert_equals: href expected "http://!\"$&'()*+,-.;=_`{}~/" but got "http://%21%22%24%26%27%28%29%2A+%2C-.%3B%3D_%60%7B%7D%7E/"
FAIL Parsing: <sc://!"$%&'()*+,-.;=_`{}~/> against <about:blank> assert_equals: host expected "%01%02%03%04%05%06%07%08%0B%0C%0E%0F%10%11%12%13%14%15%16%17%18%19%1A%1B%1C%1D%1E%1F%7F!\"$%&'()*+,-.;=_`{}~" but got ""
FAIL Parsing: <ftp://example.com%80/> against <about:blank> assert_equals: failure should set href to input expected "ftp://example.com%80/" but got "ftp://example.com%EF%BF%BD/"
FAIL Parsing: <ftp://example.com%A0/> against <about:blank> assert_equals: failure should set href to input expected "ftp://example.com%A0/" but got "ftp://example.com%EF%BF%BD/"
FAIL Parsing: <https://example.com%80/> against <about:blank> assert_equals: failure should set href to input expected "https://example.com%80/" but got "https://example.com%EF%BF%BD/"
FAIL Parsing: <https://example.com%A0/> against <about:blank> assert_equals: failure should set href to input expected "https://example.com%A0/" but got "https://example.com%EF%BF%BD/"
PASS Parsing: <ftp://%e2%98%83> against <about:blank>
PASS Parsing: <https://%e2%98%83> against <about:blank>
PASS Parsing: <http://127.0.0.1:10100/relative_import.html> against <about:blank>
PASS Parsing: <http://facebook.com/?foo=%7B%22abc%22> against <about:blank>
PASS Parsing: <https://localhost:3000/jqueryui@1.2.3> against <about:blank>
PASS Parsing: <h	t
t\rp://h	o
s\rt:9	0
0\r0/p	a
t\rh?q	u
e\rry#f	r
a\rg> against <about:blank>
PASS Parsing: <?a=b&c=d> against <http://example.org/foo/bar>
PASS Parsing: <??a=b&c=d> against <http://example.org/foo/bar>
PASS Parsing: <http:> against <http://example.org/foo/bar>
PASS Parsing: <http:> against <https://example.org/foo/bar>
PASS Parsing: <sc:> against <https://example.org/foo/bar>
PASS Parsing: <http://foo.bar/baz?qux#foobar> against <about:blank>
PASS Parsing: <http://foo.bar/baz?qux#foo"bar> against <about:blank>
PASS Parsing: <http://foo.bar/baz?qux#foo<bar> against <about:blank>
PASS Parsing: <http://foo.bar/baz?qux#foo>bar> against <about:blank>
PASS Parsing: <http://foo.bar/baz?qux#foo`bar> against <about:blank>
PASS Parsing: <http://1.2.3.4/> against <http://other.com/>
PASS Parsing: <http://1.2.3.4./> against <http://other.com/>
PASS Parsing: <http://192.168.257> against <http://other.com/>
PASS Parsing: <http://192.168.257.> against <http://other.com/>
PASS Parsing: <http://192.168.257.com> against <http://other.com/>
PASS Parsing: <http://256> against <http://other.com/>
PASS Parsing: <http://256.com> against <http://other.com/>
PASS Parsing: <http://999999999> against <http://other.com/>
PASS Parsing: <http://999999999.> against <http://other.com/>
PASS Parsing: <http://999999999.com> against <http://other.com/>
FAIL Parsing: <http://10000000000> against <http://other.com/> assert_equals: failure should set href to input expected "http://10000000000" but got "http://10000000000/"
PASS Parsing: <http://10000000000.com> against <http://other.com/>
PASS Parsing: <http://4294967295> against <http://other.com/>
FAIL Parsing: <http://4294967296> against <http://other.com/> assert_equals: failure should set href to input expected "http://4294967296" but got "http://4294967296/"
PASS Parsing: <http://0xffffffff> against <http://other.com/>
FAIL Parsing: <http://0xffffffff1> against <http://other.com/> assert_equals: failure should set href to input expected "http://0xffffffff1" but got "http://0xffffffff1/"
FAIL Parsing: <http://256.256.256.256> against <http://other.com/> assert_equals: failure should set href to input expected "http://256.256.256.256" but got "http://256.256.256.256/"
PASS Parsing: <https://0x.0x.0> against <about:blank>
PASS Parsing: <https://0x100000000/test> against <about:blank>
PASS Parsing: <https://256.0.0.1/test> against <about:blank>
FAIL Parsing: <https://%43%7C/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <asdf://%43|/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <asdf://%43%7C/> against <about:blank> assert_equals: host expected "%43%7C" but got ""
FAIL Parsing: <\\\.\Y:> against <about:blank> assert_equals: failure should set href to input expected "\\\\\\.\\Y:" but got ""
FAIL Parsing: <\\\.\y:> against <about:blank> assert_equals: failure should set href to input expected "\\\\\\.\\y:" but got ""
PASS Parsing: <http://[1:0::]> against <http://example.net/>
FAIL Parsing: <http://[0:1:2:3:4:5:6:7:8]> against <http://example.net/> assert_equals: failure should set href to input expected "http://[0:1:2:3:4:5:6:7:8]" but got "http://[0:1:2:3:4:5:6:7:8]/"
FAIL Parsing: <https://[0::0::0]> against <about:blank> assert_equals: failure should set href to input expected "https://[0::0::0]" but got "https://[0::0::0]/"
FAIL Parsing: <https://[0:.0]> against <about:blank> assert_equals: failure should set href to input expected "https://[0:.0]" but got "https://[0:.0]/"
FAIL Parsing: <https://[0:0:]> against <about:blank> assert_equals: failure should set href to input expected "https://[0:0:]" but got "https://[0:0:]/"
FAIL Parsing: <https://[0:1:2:3:4:5:6:7.0.0.0.1]> against <about:blank> assert_equals: failure should set href to input expected "https://[0:1:2:3:4:5:6:7.0.0.0.1]" but got "https://[0:1:2:3:4:5:6:7.0.0.0.1]/"
FAIL Parsing: <https://[0:1.00.0.0.0]> against <about:blank> assert_equals: failure should set href to input expected "https://[0:1.00.0.0.0]" but got "https://[0:1.00.0.0.0]/"
FAIL Parsing: <https://[0:1.290.0.0.0]> against <about:blank> assert_equals: failure should set href to input expected "https://[0:1.290.0.0.0]" but got "https://[0:1.290.0.0.0]/"
FAIL Parsing: <https://[0:1.23.23]> against <about:blank> assert_equals: failure should set href to input expected "https://[0:1.23.23]" but got "https://[0:1.23.23]/"
FAIL Parsing: <http://?> against <about:blank> assert_equals: failure should set href to input expected "http://?" but got "http:/?"
FAIL Parsing: <http://#> against <about:blank> assert_equals: failure should set href to input expected "http://#" but got "http:/#"
PASS Parsing: <http://f:4294967377/c> against <http://example.org/>
PASS Parsing: <http://f:18446744073709551697/c> against <http://example.org/>
PASS Parsing: <http://f:340282366920938463463374607431768211537/c> against <http://example.org/>
FAIL Parsing: <sc://ñ> against <about:blank> assert_equals: host expected "%C3%B1" but got ""
FAIL Parsing: <sc://ñ?x> against <about:blank> assert_equals: host expected "%C3%B1" but got ""
FAIL Parsing: <sc://ñ#x> against <about:blank> assert_equals: host expected "%C3%B1" but got ""
FAIL Parsing: <#x> against <sc://ñ> assert_equals: href expected "sc://%C3%B1#x" but got "sc://%C3%B1"
FAIL Parsing: <?x> against <sc://ñ> assert_equals: href expected "sc://%C3%B1?x" but got "sc://%C3%B1"
FAIL Parsing: <sc://?> against <about:blank> assert_equals: pathname expected "" but got "//"
FAIL Parsing: <sc://#> against <about:blank> assert_equals: pathname expected "" but got "//"
FAIL Parsing: <///> against <sc://x/> assert_equals: href expected "sc:///" but got "sc:"
FAIL Parsing: <////> against <sc://x/> assert_equals: href expected "sc:////" but got "sc:"
FAIL Parsing: <////x/> against <sc://x/> assert_equals: href expected "sc:////x/" but got "sc://x/"
FAIL Parsing: <tftp://foobar.com/someconfig;mode=netascii> against <about:blank> assert_equals: host expected "foobar.com" but got ""
FAIL Parsing: <telnet://user:pass@foobar.com:23/> against <about:blank> assert_equals: username expected "user" but got ""
FAIL Parsing: <ut2004://10.10.10.10:7777/Index.ut2> against <about:blank> assert_equals: host expected "10.10.10.10:7777" but got ""
FAIL Parsing: <redis://foo:bar@somehost:6379/0?baz=bam&qux=baz> against <about:blank> assert_equals: username expected "foo" but got ""
FAIL Parsing: <rsync://foo@host:911/sup> against <about:blank> assert_equals: username expected "foo" but got ""
FAIL Parsing: <git://github.com/foo/bar.git> against <about:blank> assert_equals: host expected "github.com" but got ""
FAIL Parsing: <irc://myserver.com:6999/channel?passwd> against <about:blank> assert_equals: host expected "myserver.com:6999" but got ""
FAIL Parsing: <dns://fw.example.org:9999/foo.bar.org?type=TXT> against <about:blank> assert_equals: host expected "fw.example.org:9999" but got ""
FAIL Parsing: <ldap://localhost:389/ou=People,o=JNDITutorial> against <about:blank> assert_equals: host expected "localhost:389" but got ""
FAIL Parsing: <git+https://github.com/foo/bar> against <about:blank> assert_equals: host expected "github.com" but got ""
PASS Parsing: <urn:ietf:rfc:2648> against <about:blank>
PASS Parsing: <tag:joe@example.org,2001:foo/bar> against <about:blank>
FAIL Parsing: <non-spec:/.//> against <about:blank> assert_equals: pathname expected "//" but got "/.//"
FAIL Parsing: <non-spec:/..//> against <about:blank> assert_equals: href expected "non-spec:/.//" but got "non-spec:/..//"
FAIL Parsing: <non-spec:/a/..//> against <about:blank> assert_equals: href expected "non-spec:/.//" but got "non-spec:/a/..//"
FAIL Parsing: <non-spec:/.//path> against <about:blank> assert_equals: pathname expected "//path" but got "/.//path"
FAIL Parsing: <non-spec:/..//path> against <about:blank> assert_equals: href expected "non-spec:/.//path" but got "non-spec:/..//path"
FAIL Parsing: <non-spec:/a/..//path> against <about:blank> assert_equals: href expected "non-spec:/.//path" but got "non-spec:/a/..//path"
FAIL Parsing: </.//path> against <non-spec:/p> assert_equals: href expected "non-spec:/.//path" but got "non-spec://path"
FAIL Parsing: </..//path> against <non-spec:/p> assert_equals: href expected "non-spec:/.//path" but got "non-spec://path"
FAIL Parsing: <..//path> against <non-spec:/p> assert_equals: href expected "non-spec:/.//path" but got "non-spec://path"
FAIL Parsing: <a/..//path> against <non-spec:/p> assert_equals: href expected "non-spec:/.//path" but got "non-spec://path"
FAIL Parsing: <> against <non-spec:/..//p> assert_equals: href expected "non-spec:/.//p" but got "non-spec:/..//p"
FAIL Parsing: <path> against <non-spec:/..//p> assert_equals: href expected "non-spec:/.//path" but got "non-spec:/..//path"
FAIL Parsing: <../path> against <non-spec:/.//p> assert_equals: href expected "non-spec:/path" but got "non-spec:/./path"
FAIL Parsing: <non-special://%E2%80%A0/> against <about:blank> assert_equals: host expected "%E2%80%A0" but got ""
FAIL Parsing: <non-special://H%4fSt/path> against <about:blank> assert_equals: host expected "H%4fSt" but got ""
FAIL Parsing: <non-special://[1:2:0:0:5:0:0:0]/> against <about:blank> assert_equals: href expected "non-special://[1:2:0:0:5::]/" but got "non-special://[1:2:0:0:5:0:0:0]/"
FAIL Parsing: <non-special://[1:2:0:0:0:0:0:3]/> against <about:blank> assert_equals: href expected "non-special://[1:2::3]/" but got "non-special://[1:2:0:0:0:0:0:3]/"
FAIL Parsing: <non-special://[1:2::3]:80/> against <about:blank> assert_equals: host expected "[1:2::3]:80" but got ""
FAIL Parsing: <non-special://[:80/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
PASS Parsing: <blob:https://example.com:443/> against <about:blank>
PASS Parsing: <blob:d3958f5c-0777-0845-9dcf-2cb28783acaf> against <about:blank>
PASS Parsing: <blob:> against <about:blank>
PASS Parsing: <http://0x7f.0.0.0x7g> against <about:blank>
PASS Parsing: <http://0X7F.0.0.0X7G> against <about:blank>
FAIL Parsing: <http://[::127.0.0.0.1]> against <about:blank> assert_equals: failure should set href to input expected "http://[::127.0.0.0.1]" but got "http://[::127.0.0.0.1]/"
PASS Parsing: <http://[0:1:0:1:0:1:0:1]> against <about:blank>
PASS Parsing: <http://[1:0:1:0:1:0:1:0]> against <about:blank>
PASS Parsing: <http://example.org/test?"> against <about:blank>
PASS Parsing: <http://example.org/test?#> against <about:blank>
PASS Parsing: <http://example.org/test?<> against <about:blank>
PASS Parsing: <http://example.org/test?>> against <about:blank>
PASS Parsing: <http://example.org/test?⌣> against <about:blank>
PASS Parsing: <http://example.org/test?%23%23> against <about:blank>
PASS Parsing: <http://example.org/test?%GH> against <about:blank>
PASS Parsing: <http://example.org/test?a#%EF> against <about:blank>
PASS Parsing: <http://example.org/test?a#%GH> against <about:blank>
FAIL Parsing: <a> against <about:blank> assert_equals: failure should set href to input expected "a" but got ""
FAIL Parsing: <a/> against <about:blank> assert_equals: failure should set href to input expected "a/" but got ""
FAIL Parsing: <a//> against <about:blank> assert_equals: failure should set href to input expected "a//" but got ""
FAIL Parsing: <test-a-colon.html> against <a:> assert_equals: failure should set href to input expected "test-a-colon.html" but got ""
FAIL Parsing: <test-a-colon-b.html> against <a:b> assert_equals: failure should set href to input expected "test-a-colon-b.html" but got ""
PASS Parsing: <test-a-colon-slash.html> against <a:/>
FAIL Parsing: <test-a-colon-slash-slash.html> against <a://> assert_equals: href expected "a:///test-a-colon-slash-slash.html" but got ""
PASS Parsing: <test-a-colon-slash-b.html> against <a:/b>
FAIL Parsing: <test-a-colon-slash-slash-b.html> against <a://b> assert_equals: href expected "a://b/test-a-colon-slash-slash-b.html" but got "a://b"
PASS Parsing: <http://example.org/test?a#b\0c> against <about:blank>
FAIL Parsing: <non-spec://example.org/test?a#b\0c> against <about:blank> assert_equals: host expected "example.org" but got ""
PASS Parsing: <non-spec:/test?a#b\0c> against <about:blank>
PASS Parsing: <a1234567890-+.:foo/bar> against <http://example.com/dir/file>
PASS Parsing: <#link> against <https://example.org/##link>
PASS Parsing: <non-special:cannot-be-a-base-url-\0~> against <about:blank>
PASS Parsing: <https://www.example.com/path{path.html?query'=query#fragment<fragment> against <about:blank>
PASS Parsing: <https://user:pass[@foo/bar> against <http://example.org>
FAIL Parsing: <foo:// !"$%&'()*+,-.;<=>@[\]^_`{|}~@host/> against <about:blank> assert_equals: href expected "foo://%20!%22$%&'()*+,-.%3B%3C%3D%3E%40%5B%5C%5D%5E_%60%7B%7C%7D~@host/" but got "foo:// !\"$%&'()*+,-.;<=>@[\\]^_`{|}~@host/"
FAIL Parsing: <wss:// !"$%&'()*+,-.;<=>@[]^_`{|}~@host/> against <about:blank> assert_equals: href expected "wss://%20!%22$%&'()*+,-.%3B%3C%3D%3E%40%5B%5D%5E_%60%7B%7C%7D~@host/" but got "wss://%20!%22$%&%27()*+,-.%3B%3C%3D%3E%40%5B%5D%5E_%60%7B%7C%7D~@host/"
FAIL Parsing: <foo://joe: !"$%&'()*+,-.:;<=>@[\]^_`{|}~@host/> against <about:blank> assert_equals: href expected "foo://joe:%20!%22$%&'()*+,-.%3A%3B%3C%3D%3E%40%5B%5C%5D%5E_%60%7B%7C%7D~@host/" but got "foo://joe: !\"$%&'()*+,-.:;<=>@[\\]^_`{|}~@host/"
FAIL Parsing: <wss://joe: !"$%&'()*+,-.:;<=>@[]^_`{|}~@host/> against <about:blank> assert_equals: href expected "wss://joe:%20!%22$%&'()*+,-.%3A%3B%3C%3D%3E%40%5B%5D%5E_%60%7B%7C%7D~@host/" but got "wss://joe:%20!%22$%&%27()*+,-.%3A%3B%3C%3D%3E%40%5B%5D%5E_%60%7B%7C%7D~@host/"
FAIL Parsing: <foo://!"$%&'()*+,-.;=_`{}~/> against <about:blank> assert_equals: host expected "!\"$%&'()*+,-.;=_`{}~" but got ""
FAIL Parsing: <wss://!"$&'()*+,-.;=_`{}~/> against <about:blank> assert_equals: href expected "wss://!\"$&'()*+,-.;=_`{}~/" but got "wss://%21%22%24%26%27%28%29%2A+%2C-.%3B%3D_%60%7B%7D%7E/"
FAIL Parsing: <foo://host/ !"$%&'()*+,-./:;<=>@[\]^_`{|}~> against <about:blank> assert_equals: href expected "foo://host/%20!%22$%&'()*+,-./:;%3C=%3E@[\\]^_%60%7B|%7D~" but got "foo://host/ !\"$%&'()*+,-./:;<=>@[\\]^_`{|}~"
FAIL Parsing: <wss://host/ !"$%&'()*+,-./:;<=>@[\]^_`{|}~> against <about:blank> assert_equals: href expected "wss://host/%20!%22$%&'()*+,-./:;%3C=%3E@[/]^_%60%7B|%7D~" but got "wss://host/%20!%22$%&'()*+,-./:;%3C=%3E@[/]%5E_%60%7B%7C%7D~"
FAIL Parsing: <foo://host/dir/? !"$%&'()*+,-./:;<=>?@[\]^_`{|}~> against <about:blank> assert_equals: href expected "foo://host/dir/?%20!%22$%&'()*+,-./:;%3C=%3E?@[\\]^_`{|}~" but got "foo://host/dir/?%20!%22$%&%27()*+,-./:;%3C=%3E?@[\\]^_`{|}~"
PASS Parsing: <wss://host/dir/? !"$%&'()*+,-./:;<=>?@[\]^_`{|}~> against <about:blank>
FAIL Parsing: <foo://host/dir/# !"#$%&'()*+,-./:;<=>?@[\]^_`{|}~> against <about:blank> assert_equals: host expected "host" but got ""
PASS Parsing: <wss://host/dir/# !"#$%&'()*+,-./:;<=>?@[\]^_`{|}~> against <about:blank>
FAIL Parsing: <abc:rootless> against <abc://host/path> assert_equals: href expected "abc:rootless" but got "abc://host/rootless"
FAIL Parsing: <abc:rootless> against <abc:/path> assert_equals: href expected "abc:rootless" but got "abc:/rootless"
PASS Parsing: <abc:rootless> against <abc:path>
FAIL Parsing: <abc:/rooted> against <abc://host/path> assert_equals: href expected "abc:/rooted" but got "abc://host/rooted"
FAIL Parsing: <http://1.2.3.4.5> against <http://other.com/> assert_equals: failure should set href to input expected "http://1.2.3.4.5" but got "http://1.2.3.4.5/"
FAIL Parsing: <http://1.2.3.4.5.> against <http://other.com/> assert_equals: failure should set href to input expected "http://1.2.3.4.5." but got "http://1.2.3.4.5./"
PASS Parsing: <http://0..0x300/> against <about:blank>
PASS Parsing: <http://0..0x300./> against <about:blank>
FAIL Parsing: <http://256.256.256.256.256> against <http://other.com/> assert_equals: failure should set href to input expected "http://256.256.256.256.256" but got "http://256.256.256.256.256/"
FAIL Parsing: <http://256.256.256.256.256.> against <http://other.com/> assert_equals: failure should set href to input expected "http://256.256.256.256.256." but got "http://256.256.256.256.256./"
FAIL Parsing: <http://1.2.3.08> against <about:blank> assert_equals: failure should set href to input expected "http://1.2.3.08" but got "http://1.2.3.08/"
FAIL Parsing: <http://1.2.3.08.> against <about:blank> assert_equals: failure should set href to input expected "http://1.2.3.08." but got "http://1.2.3.08./"
FAIL Parsing: <http://1.2.3.09> against <about:blank> assert_equals: failure should set href to input expected "http://1.2.3.09" but got "http://1.2.3.09/"
FAIL Parsing: <http://09.2.3.4> against <about:blank> assert_equals: failure should set href to input expected "http://09.2.3.4" but got "http://09.2.3.4/"
FAIL Parsing: <http://09.2.3.4.> against <about:blank> assert_equals: failure should set href to input expected "http://09.2.3.4." but got "http://09.2.3.4./"
FAIL Parsing: <http://01.2.3.4.5> against <about:blank> assert_equals: failure should set href to input expected "http://01.2.3.4.5" but got "http://01.2.3.4.5/"
FAIL Parsing: <http://01.2.3.4.5.> against <about:blank> assert_equals: failure should set href to input expected "http://01.2.3.4.5." but got "http://01.2.3.4.5./"
FAIL Parsing: <http://0x100.2.3.4> against <about:blank> assert_equals: failure should set href to input expected "http://0x100.2.3.4" but got "http://0x100.2.3.4/"
FAIL Parsing: <http://0x100.2.3.4.> against <about:blank> assert_equals: failure should set href to input expected "http://0x100.2.3.4." but got "http://0x100.2.3.4./"
FAIL Parsing: <http://0x1.2.3.4.5> against <about:blank> assert_equals: failure should set href to input expected "http://0x1.2.3.4.5" but got "http://0x1.2.3.4.5/"
FAIL Parsing: <http://0x1.2.3.4.5.> against <about:blank> assert_equals: failure should set href to input expected "http://0x1.2.3.4.5." but got "http://0x1.2.3.4.5./"
FAIL Parsing: <http://foo.1.2.3.4> against <about:blank> assert_equals: failure should set href to input expected "http://foo.1.2.3.4" but got "http://foo.1.2.3.4/"
FAIL Parsing: <http://foo.1.2.3.4.> against <about:blank> assert_equals: failure should set href to input expected "http://foo.1.2.3.4." but got "http://foo.1.2.3.4./"
FAIL Parsing: <http://foo.2.3.4> against <about:blank> assert_equals: failure should set href to input expected "http://foo.2.3.4" but got "http://foo.2.3.4/"
FAIL Parsing: <http://foo.2.3.4.> against <about:blank> assert_equals: failure should set href to input expected "http://foo.2.3.4." but got "http://foo.2.3.4./"
FAIL Parsing: <http://foo.09> against <about:blank> assert_equals: failure should set href to input expected "http://foo.09" but got "http://foo.09/"
FAIL Parsing: <http://foo.09.> against <about:blank> assert_equals: failure should set href to input expected "http://foo.09." but got "http://foo.09./"
FAIL Parsing: <http://foo.0x4> against <about:blank> assert_equals: failure should set href to input expected "http://foo.0x4" but got "http://foo.0x4/"
FAIL Parsing: <http://foo.0x4.> against <about:blank> assert_equals: failure should set href to input expected "http://foo.0x4." but got "http://foo.0x4./"
PASS Parsing: <http://foo.09..> against <about:blank>
PASS Parsing: <http://0999999999999999999/> against <about:blank>
FAIL Parsing: <http://foo.0x> against <about:blank> assert_equals: failure should set href to input expected "http://foo.0x" but got "http://foo.0x/"
FAIL Parsing: <http://foo.0XFfFfFfFfFfFfFfFfFfAcE123> against <about:blank> assert_equals: failure should set href to input expected "http://foo.0XFfFfFfFfFfFfFfFfFfAcE123" but got "http://foo.0xfffffffffffffffffface123/"
FAIL Parsing: <http://💩.123/> against <about:blank> assert_equals: failure should set href to input expected "http://💩.123/" but got "http://xn--ls8h.123/"
FAIL Parsing: <https://\0y> against <about:blank> assert_equals: failure should set href to input expected "https://\0y" but got "https://%00y/"
FAIL Parsing: <https://x/\0y> against <about:blank> assert_equals: protocol expected "https:" but got ":"
PASS Parsing: <https://x/?\0y> against <about:blank>
PASS Parsing: <https://x/?#\0y> against <about:blank>
FAIL Parsing: <https://￿y> against <about:blank> assert_equals: failure should set href to input expected "https://\uffffy" but got "https://%EF%BF%BDy/"
FAIL Parsing: <https://x/￿y> against <about:blank> assert_equals: href expected "https://x/%EF%BF%BFy" but got "https://x/%EF%BF%BDy"
FAIL Parsing: <https://x/?￿y> against <about:blank> assert_equals: href expected "https://x/?%EF%BF%BFy" but got "https://x/?%EF%BF%BDy"
FAIL Parsing: <https://x/?#￿y> against <about:blank> assert_equals: href expected "https://x/?#%EF%BF%BFy" but got "https://x/?#%EF%BF%BDy"
PASS Parsing: <non-special:\0y> against <about:blank>
PASS Parsing: <non-special:x/\0y> against <about:blank>
PASS Parsing: <non-special:x/?\0y> against <about:blank>
PASS Parsing: <non-special:x/?#\0y> against <about:blank>
FAIL Parsing: <non-special:￿y> against <about:blank> assert_equals: href expected "non-special:%EF%BF%BFy" but got "non-special:%EF%BF%BDy"
FAIL Parsing: <non-special:x/￿y> against <about:blank> assert_equals: href expected "non-special:x/%EF%BF%BFy" but got "non-special:x/%EF%BF%BDy"
FAIL Parsing: <non-special:x/?￿y> against <about:blank> assert_equals: href expected "non-special:x/?%EF%BF%BFy" but got "non-special:x/?%EF%BF%BDy"
FAIL Parsing: <non-special:x/?#￿y> against <about:blank> assert_equals: href expected "non-special:x/?#%EF%BF%BFy" but got "non-special:x/?#%EF%BF%BDy"
PASS Parsing: <> against <about:blank>
PASS Parsing: <https://example.com/"quoted"> against <about:blank>
PASS Parsing: <https://a%C2%ADb/> against <about:blank>
FAIL Parsing: <https://­/> against <about:blank> assert_equals: failure should set href to input expected "https://­/" but got "https://%C2%AD/"
PASS Parsing: <https://%C2%AD/> against <about:blank>
FAIL Parsing: <https://xn--/> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <data://example.com:8080/pathname?search#hash> against <about:blank> assert_equals: host expected "example.com:8080" but got ""
FAIL Parsing: <data:///test> against <about:blank> assert_equals: pathname expected "/test" but got "///test"
FAIL Parsing: <data://test/a/../b> against <about:blank> assert_equals: href expected "data://test/b" but got "data://test/a/../b"
FAIL Parsing: <data://:443> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <data://test:test> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <data://[:1]> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <intent://example.com:8080/pathname?search#hash> against <about:blank> assert_equals: host expected "example.com:8080" but got ""
FAIL Parsing: <intent:///test> against <about:blank> assert_equals: pathname expected "/test" but got "///test"
FAIL Parsing: <intent://test/a/../b> against <about:blank> assert_equals: href expected "intent://test/b" but got "intent://test/a/../b"
FAIL Parsing: <intent://:443> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <intent://test:test> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <intent://[:1]> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <urn://example.com:8080/pathname?search#hash> against <about:blank> assert_equals: host expected "example.com:8080" but got ""
FAIL Parsing: <urn:///test> against <about:blank> assert_equals: pathname expected "/test" but got "///test"
FAIL Parsing: <urn://test/a/../b> against <about:blank> assert_equals: href expected "urn://test/b" but got "urn://test/a/../b"
FAIL Parsing: <urn://:443> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <urn://test:test> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <urn://[:1]> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <turn://example.com:8080/pathname?search#hash> against <about:blank> assert_equals: host expected "example.com:8080" but got ""
FAIL Parsing: <turn:///test> against <about:blank> assert_equals: pathname expected "/test" but got "///test"
FAIL Parsing: <turn://test/a/../b> against <about:blank> assert_equals: href expected "turn://test/b" but got "turn://test/a/../b"
FAIL Parsing: <turn://:443> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <turn://test:test> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <turn://[:1]> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <stun://example.com:8080/pathname?search#hash> against <about:blank> assert_equals: host expected "example.com:8080" but got ""
FAIL Parsing: <stun:///test> against <about:blank> assert_equals: pathname expected "/test" but got "///test"
FAIL Parsing: <stun://test/a/../b> against <about:blank> assert_equals: href expected "stun://test/b" but got "stun://test/a/../b"
FAIL Parsing: <stun://:443> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <stun://test:test> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
FAIL Parsing: <stun://[:1]> against <about:blank> assert_unreached: Expected URL to fail parsing Reached unreachable code
Harness: the test ran to completion.
